fix: 他の記事ページでも使えるデザインを共通コンポーネント化
