+++ frontend/components/common/article/index.tsx
@@ -1,0 +1,2 @@
+export { ArticleOutlineFrame } from "./ArticleOutlineFrame";
+export { ArticleTitle } from "./ArticleTitle";
